requirements.txt
# --- Core AI / LLM ---
llama-cpp-python>=0.2.90

# --- Utilities ---
numpy>=1.26
typing-extensions>=4.8

# --- Optional (for later upgrades) ---
# sentence-transformers>=2.6.0
# faiss-cpu>=1.8.0
